docker/adapters/nvclip-adapters/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0
python-multipart==0.0.6
pillow==10.0.1